requirements.txt
huggingface_hub==0.29.3
gradio~=5.22.0
huggingface-hub~=0.29.3
litellm~=1.63.14
llama-index-core~=0.12.25
llama-index-vector-stores-chroma~=0.4.1
llama-index-embeddings-cohere~=0.4.0
llama-index-embeddings-huggingface~=0.5.2
cohere~=5.14.0
numpy~=2.2.4
sentence-transformers~=3.4.1
tqdm~=4.67.1
chromadb~=0.6.3
torch~=2.6.0
requests~=2.32.3
groq~=0.20.0
openai~=1.68.2
elevenlabs~=1.54.0